refactor vct_lazy module
