+++ src/mate/vct_lazy.rs
@@ -1,3 +1,10 @@
+/*
+Lazy VCT solver is earlier experimental implementation and not maintained well.
+The idea was inspired by following paper.
+
+長井歩. "難解な必至問題を解くアルゴリズムとその実装." ゲームプログラミングワークショップ 2011 論文集 2011.6 (2011): 1-8.
+*/
+
 mod field;
 mod generator;
 mod helper;
@@ -8,6 +15,5 @@ mod solver;
 mod state;
 mod traverser;
 
-pub use solver::LazyDFPNSSolver;
-pub use solver::LazySolver;
-pub use state::VCTState;
+pub use solver::LazyVCTSolver;
+pub use state::LazyVCTState;
